Drop trait since we no longer need it
