--- crates/cow-amm/src/implementations/mod.rs
@@ -1,1 +1,0 @@
-pub(crate) mod standalone;
